Cleans up PR 860
